Making interface for INeo4jInstaller and extracting common code into helper classes. This will be used for creating a process based Neo4jInstaller that will be platform independed
